python_requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
streamlit==1.3.1

